% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercensal_survival.R
\name{intercensal_surv_Preston_logit_var_r}
\alias{intercensal_surv_Preston_logit_var_r}
\title{Estimate life expectancy at age 5 (\code{e(5)}) using cohort relationship between two censuses, based on r-variable methods.}
\usage{
intercensal_surv_Preston_logit_var_r(
  c1,
  c2,
  date1,
  date2,
  age1 = 1:length(c1) - 1,
  age2 = 1:length(c2) - 1,
  sex = "f",
  mlt_family = "CD_West",
  mlt_e0 = 60,
  ages_fit = NULL,
  q0_5 = NULL,
  verbose = FALSE
)
}
\arguments{
\item{c1}{numeric vector. Population counts in five-age groups, from first census with an exact reference date.}

\item{c2}{numeric vector. Population counts in five-age groups, from second census with an exact reference date.}

\item{date1}{Either a \code{Date} class object or an unambiguous character string in the format \code{"YYYY-MM-DD"}. Reference date for first census.}

\item{date2}{Either a \code{Date} class object or an unambiguous character string in the format \code{"YYYY-MM-DD"}. Reference date for second census.}

\item{age1}{integer vector. Lower bound of age groups from first census. Last age is assumed OAG.}

\item{age2}{integer vector. Lower bound of age groups from second census. Last age is assumed OAG.}

\item{sex}{character string. Either \code{"m"} for males, \code{"f"} for females (default).}

\item{mlt_family}{character string. Model life table family: \code{"Chilean"}, \code{"East"}, \code{"Far_East_Asian"}, \code{"General"}, \code{"Latin"}, \code{"North"}, \code{"South"}, \code{"South_Asian"} or \code{"West"} (default).}

\item{ages_fit}{integer vector. Ages to be considered when calculating the median of implied level by age. By default 10 to 70, but depends data and method.}

\item{q05}{numeric. Probability of dying from birth to age 5.}
}
\description{
Method developed by Preston (2001, section 11.5.2)
}
