% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lt5_wpp24_brazil}
\alias{lt5_wpp24_brazil}
\title{WPP 2024 five-year life tables for Brazil}
\format{
A \code{data.frame} with 29 columns:
\describe{
\item{SortOrder, LocID, VarID, Time, SexID, AgeGrpStart, AgeGrpSpan}{Integer identifiers and time or age fields.}
\item{Notes, ISO3_code, ISO2_code, LocTypeName, Location, Variant, Sex, AgeGrp}{Character metadata fields.}
\item{MidPeriod, mx, qx, px, lx, dx, Lx, Sx, Tx, ex, ax}{Numeric life table values.}
}
}
\source{
United Nations World Population Prospects 2024.
}
\usage{
lt5_wpp24_brazil
}
\description{
Brazil Country-level five-year life table values from the 2024 World Population
Prospects extract.
}
\keyword{datasets}
