% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mac_wpp22}
\alias{mac_wpp22}
\title{WPP 2022 mean age of childbearing}
\format{
A tibble with 37,488 rows and 4 columns:
\describe{
\item{LocID}{Integer location identifier.}
\item{Year}{Integer calendar year.}
\item{parent}{Parent sex category.}
\item{mac}{Mean age of childbearing.}
}
}
\source{
United Nations World Population Prospects 2022.
}
\usage{
mac_wpp22
}
\description{
Mean age of childbearing by location, year, and parent sex.
}
\keyword{datasets}
