% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mods.r}
\alias{mods.r}
\title{Spectrum mortality models}
\format{
A named list of length 2 with elements \code{female} and \code{male}.
}
\source{
Spectrum mortality model files.
}
\usage{
mods.r
}
\description{
Fitted Spectrum mortality model components for female and male mortality
estimation.
}
\keyword{datasets}
