% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Orphanhood_DB}
\alias{Orphanhood_DB}
\title{Orphanhood database}
\format{
A \code{data.table} and \code{data.frame} with 59,937 rows and 93 columns.
Important fields include \code{LocName}, \code{LocID}, \code{DataCatalogName},
\code{ReferenceYearStart}, \code{ReferenceYearEnd}, \code{SexName}, \code{AgeStart}, \code{AgeEnd},
\code{PeriodStart}, \code{PeriodEnd}, and \code{DataValue}.
}
\source{
United Nations structured demographic data extract.
}
\usage{
Orphanhood_DB
}
\description{
Census and survey records of parental survival by respondent age, sex,
location, source, period, and related metadata.
}
\keyword{datasets}
