% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mods.r}
\alias{mods.r}
\title{Spectrum model
A list with Spectrum model for each sex (updated on ...)}
\format{
An object of class \code{list} of length 2.
}
\usage{
mods.r
}
\description{
Spectrum model
A list with Spectrum model for each sex (updated on ...)
}
\keyword{datasets}
