% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orphanhood.R
\name{orphanhood_two}
\alias{orphanhood_two}
\title{Estimate adult mortality with two survey/census observations from orphanhood data.}
\usage{
orphanhood_two(
  prop1_not_orphan,
  prop2_not_orphan,
  age1,
  age2,
  date1,
  date2,
  maternal = TRUE,
  method = "timæus",
  mac1 = NULL,
  mac2 = NULL,
  mlt_data_input = NULL,
  mlt_family = NULL,
  brass_logit = FALSE,
  brass_logit_e0 = 60,
  HIV_prev = NULL,
  HIV_art = NULL,
  HIV_5q0 = NULL,
  e0_accept = c(20, 100),
  verbose = TRUE
)
}
\arguments{
\item{prop1_not_orphan}{numeric vector. Population proportion of respondents with mother/father alive, by age groups, from source 1.}

\item{prop2_not_orphan}{numeric vector. Population proportion of respondents with mother/father alive, by age groups, from source 2.}

\item{age1}{integer vector. Lower bound of age groups from first census. Last age is assumed as lower age open age group.}

\item{age2}{Same than \code{age1} but for the second source.}

\item{date1}{Either a \code{Date} class object or an unambiguous character string in the format \code{"YYYY-MM-DD"}. Reference date for the source 1 (typically census or survey).}

\item{date2}{Same than \code{date1} but for the second source.}

\item{maternal}{logical. Either \code{TRUE} for maternal orphanhood (default) or \code{FALSE} for paternal orphanhood.}

\item{method}{character. Choose between \code{timæus} (default) (Timæus, 1991) or \code{preston_chen} using r-variable method (Preston, 2001, p. 239).}

\item{mac1}{numeric. Mean age at childbearing for mothers/fathers from source 1.}

\item{mac2}{Same than \code{mac1} but for the second source.}

\item{mlt_data_input}{data.frame. If is \code{NULL} then model life tables is used, available from \code{Morcast} package. But specific pattern can be included (\code{l_x} must be included see examples).}

\item{mlt_family}{character. Options: "CD_East", "CD_North", "CD_South", "CD_West", "UN_Chilean", "UN_Far_Eastern", "UN_General", "UN_Latin_American", "UN_South_Asian". If \code{NULL} returns for all.}

\item{brass_logit}{logical. Doing a level smoothing with 1-parameter logit Brass.}

\item{brass_logit_e0}{numeric. Life expectancy at birth when \code{brass_logit} is \code{TRUE}.}

\item{e0_accept}{integer vector. Range acceptable when calculating the median of implied level by age (avoid non-possible extrapolations). By deafult between 20 and 100.}
}
\description{
Follows IUSSP(2012) template implementation of Timæus (1992) for estimate adult mortality using two observations.
The other option is using r-variable method (Preston, 2001, p. 239).
}
\examples{
\dontrun{
# Example from IUUSP Tools for Demographic Estimation
# Orhpanhood 2 census
# Paternal orphanhood - Salomons 1987 - 1999 ("AM_Orphanhood_TwoCensus - Solomons_2.xlsx")
salomons_1986_1999_paternal <- orphanhood_two(
  prop1_not_orphan = c(0.9644 ,0.9340 ,0.8862 ,0.8088 ,0.7060 ,0.5812 ,0.4620 ,0.3317),
  prop2_not_orphan = c(0.9710 ,0.9454 ,0.9071 ,0.8493 ,0.7717 ,0.6568 ,0.5375 ,0.3891),
  age1 = seq(5, 40, 5), age2 = seq(5, 40, 5),
  date1 = 1986.9, date2 = 1999.9,
  maternal = FALSE,
  mac1 = 34.08, mac2 = 34.08,
  mlt_data_input = standard_iusssp,
  brass_logit = TRUE)$adult_mort
rr_45q15 <- c(0.175, 0.185,0.187,0.187)
round((salomons_1986_1999_paternal$q15_45/rr_45q15)-1,2)
# Preston_chen method
# table 11.3 Preston (2001) r-var
panama_1977_1980 <- orphanhood_two(
  prop1_not_orphan = c(.996,.9844,.9753,.9556,.9195,.8908,.8157,.7618,.6166),
  prop2_not_orphan = c(.9949,.9903,.9794,.9608,.9316,.8967,.8450,.7746,.6781),
  age1 = seq(0, 40, 5),
  age2 = seq(0, 40, 5),
  date1 = "1977-07-01",
  date2 = "1980-05-11",
  maternal = TRUE,
  mac1 = 27,
  mac2 = 27,
  method = "preston_chen",
  mlt_data_input = NULL,
  mlt_family = "CD_West",
  brass_logit = FALSE,
  brass_logit_e0 = 60)
rr_levels <- c(22.4, 23, 22.7, 22.6, 22.9, 23, 23.5, 23.9)
}
}
